Fixed IE8 property name bug
Reserved identifiers are not converted to strings when used as property names in some cases in IE8.
So we need to change stuff like foo.null to foo[null].
I have updated null and enum as they currently cause an issue in IE8. It might make sense to change other.